requirements.txt
# Discord Bot
discord.py==2.3.2
discord-py-interactions==5.11.0

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1

# Database (optional)
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
black==23.12.0
flake8==6.1.0
